Fixing Relational Data - Relational Data Validation (Part - 3)

+++ app/Http/Controllers/Admin/IndustryTypeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Company;
 use App\Models\IndustryType;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -84,7 +85,12 @@ class IndustryTypeController extends Controller
      */
     public function destroy(string $id)
     {
-        
+        $company = Company::where('industry_type_id', $id)->exists();
+
+        if($company) {
+            return response(['message' => 'This item is already been used can\'t delete!'], 500);
+        }
+
         try {
             IndustryType::findOrFail($id)->delete();
             Notify::deletedNotification();
