+++ app/Http/Controllers/Admin/LanguageController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\CandidateLanguage;
 use App\Models\Language;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -85,6 +86,12 @@ class LanguageController extends Controller
      */
     public function destroy(string $id) : Response
     {
+        $Language = CandidateLanguage::where('language_id', $id)->exists();
+
+        if($Language) {
+            return response(['message' => 'This item is already been used can\'t delete!'], 500);
+        }
+
         try {
             Language::findOrFail($id)->delete();
             Notify::deletedNotification();
